+++ Fody/Properties/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 ﻿using System.Runtime.CompilerServices;
 
-[assembly:InternalsVisibleTo("Tests")]
+[assembly:InternalsVisibleTo("Stamp.Fody.Tests")]
