+++ Tests/packages.config
@@ -5,4 +5,5 @@
   <package id="LibGit2Sharp" version="0.22.0" targetFramework="net40" />
   <package id="LibGit2Sharp.NativeBinaries" version="1.0.129" targetFramework="net40" />
   <package id="NUnit" version="3.10.1" targetFramework="net452" />
+  <package id="System.ValueTuple" version="4.3.0" targetFramework="net46" />
 </packages>
